move ScopedResourcesBuilder
